+++ src/test/endpoints/study/exports.test.ts
@@ -131,7 +131,7 @@ describe("Exports", () => {
     expect(downloadedExport).toBeDefined();
   });
 
-  test("anonymous participant can be generated", async () => {
+  test("anonymous participant without subdomain can be generated", async () => {
     const response =
       await testClient.study.recruitment.generateAnonymousAccounts({
         studyId: study.studyId.stringRepresentation,
@@ -139,6 +139,31 @@ describe("Exports", () => {
         expirationSeconds: 3600,
         clientId: "studies-app",
         redirectUri: "carp-studies:/callback",
+        participantRoleName: "Participant",
+      });
+
+    expect(response).toBeDefined();
+    await new Promise((resolve) => {
+      setTimeout(resolve, 3000);
+    });
+
+    const exports = await testClient.study.exports.getAll({
+      studyId: study.studyId.stringRepresentation,
+    });
+    expect(
+      exports.find((e) => e.type === "ANONYMOUS_PARTICIPANTS").status,
+    ).toBe("AVAILABLE");
+  });
+
+  test("anonymous participant with subdomain can be generated", async () => {
+    const response =
+      await testClient.study.recruitment.generateAnonymousAccounts({
+        studyId: study.studyId.stringRepresentation,
+        amountOfAccounts: 1,
+        expirationSeconds: 3600,
+        clientId: "studies-app",
+        redirectUri: "https://csa.dev.carp.dk/anonyomous",
+        subdomain: null,
         participantRoleName: "Participant",
       });